add effects! and handler! for grouping effects
when you group effects like this, the handler macro uses the same closure for
handling all of them, so this closure can hold a mutable borrow and the borrow
can be used in handling all of the effects.

it also means that an effectful function can use all of the effects by naming
only the group.